Split coil-network-ktor into ktor2 and ktor3 artifacts.
