+++ buildSrc/src/main/kotlin/coil3/utils.kt
@@ -9,7 +9,8 @@ val publicModules = setOf(
     "coil-compose",
     "coil-compose-core",
     "coil-network-core",
-    "coil-network-ktor",
+    "coil-network-ktor2",
+    "coil-network-ktor3",
     "coil-network-okhttp",
     "coil-gif",
     "coil-svg",
