+++ coil-core/build.gradle.kts
@@ -2,6 +2,7 @@ import coil3.addAllMultiplatformTargets
 import coil3.androidInstrumentedTest
 import coil3.androidLibrary
 import coil3.androidUnitTest
+import coil3.createSkikoWasmJsRuntimeDependency
 
 plugins {
     id("com.android.library")
@@ -12,6 +13,7 @@ plugins {
 }
 
 addAllMultiplatformTargets()
+createSkikoWasmJsRuntimeDependency()
 androidLibrary(name = "coil3.core")
 
 kotlin {
