--- coil-network-ktor/build.gradle.kts
@@ -1,38 +1,0 @@
-import coil3.addAllMultiplatformTargets
-import coil3.androidLibrary
-import coil3.androidUnitTest
-import coil3.applyKtorWasmWorkaround
-
-plugins {
-    id("com.android.library")
-    id("kotlin-multiplatform")
-    id("org.jetbrains.kotlinx.atomicfu")
-}
-
-addAllMultiplatformTargets()
-androidLibrary(name = "coil3.network.ktor")
-applyKtorWasmWorkaround(libs.versions.ktor.beta.get())
-
-kotlin {
-    sourceSets {
-        commonMain {
-            dependencies {
-                api(projects.coilCore)
-                api(projects.coilNetworkCore)
-                api(libs.ktor.core)
-            }
-        }
-        commonTest {
-            dependencies {
-                implementation(projects.internal.testUtils)
-                implementation(libs.bundles.test.common)
-                implementation(libs.ktor.mock)
-            }
-        }
-        androidUnitTest {
-            dependencies {
-                implementation(libs.bundles.test.jvm)
-            }
-        }
-    }
-}
