--- coil-network-ktor/gradle.properties
@@ -1,3 +1,0 @@
-POM_ARTIFACT_ID=coil-network-ktor
-POM_NAME=coil-network-ktor
-POM_PACKAGING=aar
