--- coil-network-ktor/src/jsMain/kotlin/coil3/network/ktor/internal/KtorNetworkFetcherServiceLoaderTarget.js.kt
@@ -1,12 +1,0 @@
-package coil3.network.ktor.internal
-
-import coil3.annotation.InternalCoilApi
-import coil3.util.ServiceLoaderComponentRegistry
-
-@Suppress("DEPRECATION")
-@OptIn(ExperimentalStdlibApi::class, ExperimentalJsExport::class)
-@EagerInitialization
-@JsExport
-@InternalCoilApi
-@Deprecated("", level = DeprecationLevel.HIDDEN)
-val initHook: Any = ServiceLoaderComponentRegistry.register(KtorNetworkFetcherServiceLoaderTarget())
