--- coil-network-ktor/src/jvmCommonMain/resources/META-INF/services/coil3.util.FetcherServiceLoaderTarget
@@ -1,1 +1,0 @@
-coil3.network.ktor.internal.KtorNetworkFetcherServiceLoaderTarget
