--- coil-network-ktor/src/wasmJsMain/kotlin/coil3/network/ktor/internal/KtorNetworkFetcherServiceLoaderTarget.wasmJs.kt
@@ -1,11 +1,0 @@
-package coil3.network.ktor.internal
-
-import coil3.annotation.InternalCoilApi
-import coil3.util.ServiceLoaderComponentRegistry
-
-@Suppress("DEPRECATION")
-@OptIn(ExperimentalStdlibApi::class)
-@EagerInitialization
-@InternalCoilApi
-@Deprecated("", level = DeprecationLevel.HIDDEN)
-val initHook: Any = ServiceLoaderComponentRegistry.register(KtorNetworkFetcherServiceLoaderTarget())
