+++ coil-network-ktor2/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=coil-network-ktor2
+POM_NAME=coil-network-ktor2
+POM_PACKAGING=aar
