+++ coil-network-ktor2/src/commonTest/kotlin/coil3/network/ktor2/internal/KtorNetworkFetcherServiceLoaderTest.kt
@@ -1,0 +1,15 @@
+package coil3.network.ktor2.internal
+
+import coil3.network.NetworkFetcher
+import coil3.util.ServiceLoaderComponentRegistry
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+class KtorNetworkFetcherServiceLoaderTest {
+
+    @Test
+    fun serviceLoaderFindsFetcher() {
+        val fetchers = ServiceLoaderComponentRegistry.fetchers
+        assertTrue(fetchers.any { it.factory() is NetworkFetcher.Factory })
+    }
+}
