+++ coil-network-ktor2/src/jsMain/kotlin/coil3/network/ktor2/internal/KtorNetworkFetcherServiceLoaderTarget.js.kt
@@ -1,0 +1,12 @@
+package coil3.network.ktor2.internal
+
+import coil3.annotation.InternalCoilApi
+import coil3.util.ServiceLoaderComponentRegistry
+
+@Suppress("DEPRECATION")
+@OptIn(ExperimentalStdlibApi::class, ExperimentalJsExport::class)
+@EagerInitialization
+@JsExport
+@InternalCoilApi
+@Deprecated("", level = DeprecationLevel.HIDDEN)
+val initHook: Any = ServiceLoaderComponentRegistry.register(KtorNetworkFetcherServiceLoaderTarget())
