+++ coil-network-ktor2/src/nativeMain/kotlin/coil3/network/ktor2/internal/KtorNetworkFetcherServiceLoaderTarget.native.kt
@@ -1,0 +1,11 @@
+package coil3.network.ktor2.internal
+
+import coil3.annotation.InternalCoilApi
+import coil3.util.ServiceLoaderComponentRegistry
+
+@Suppress("DEPRECATION")
+@OptIn(ExperimentalStdlibApi::class)
+@EagerInitialization
+@InternalCoilApi
+@Deprecated("", level = DeprecationLevel.HIDDEN)
+val initHook: Any = ServiceLoaderComponentRegistry.register(KtorNetworkFetcherServiceLoaderTarget())
