+++ coil-network-ktor3/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=coil-network-ktor3
+POM_NAME=coil-network-ktor3
+POM_PACKAGING=aar
