+++ coil-network-ktor3/src/jvmCommonMain/resources/META-INF/services/coil3.util.FetcherServiceLoaderTarget
@@ -1,0 +1,1 @@
+coil3.network.ktor3.internal.KtorNetworkFetcherServiceLoaderTarget
