+++ coil-network-ktor3/src/wasmJsMain/kotlin/coil3/network/ktor3/internal/KtorNetworkFetcherServiceLoaderTarget.wasmJs.kt
@@ -1,0 +1,11 @@
+package coil3.network.ktor3.internal
+
+import coil3.annotation.InternalCoilApi
+import coil3.util.ServiceLoaderComponentRegistry
+
+@Suppress("DEPRECATION")
+@OptIn(ExperimentalStdlibApi::class)
+@EagerInitialization
+@InternalCoilApi
+@Deprecated("", level = DeprecationLevel.HIDDEN)
+val initHook: Any = ServiceLoaderComponentRegistry.register(KtorNetworkFetcherServiceLoaderTarget())
