+++ gradle/libs.versions.toml
@@ -8,8 +8,8 @@ coroutines = "1.8.1"
 jetbrains-compose = "1.6.11"
 kotlin = "2.0.0"
 ktlint = "1.1.0"
-ktor = "2.3.12"
-ktor-beta = "3.0.0-beta-2"
+ktor2 = "2.3.12"
+ktor3 = "3.0.0-beta-2"
 okhttp = "4.12.0"
 okio = "3.9.0"
 paparazzi = "1.3.4"
@@ -76,10 +76,13 @@ kotlin-test-junit = { module = "org.jetbrains.kotlin:kotlin-test-junit" }
 kotlinx-datetime = "org.jetbrains.kotlinx:kotlinx-datetime:0.6.0"
 kotlinx-serialization-json = "org.jetbrains.kotlinx:kotlinx-serialization-json:1.7.1"
 
-ktor-core = { module = "io.ktor:ktor-client-core", version.ref = "ktor" }
-ktor-engine-darwin = { module = "io.ktor:ktor-client-darwin", version.ref = "ktor" }
-ktor-engine-js = { module = "io.ktor:ktor-client-js", version.ref = "ktor" }
-ktor-mock = { module = "io.ktor:ktor-client-mock", version.ref = "ktor" }
+ktor2-core = { module = "io.ktor:ktor-client-core", version.ref = "ktor2" }
+ktor2-mock = { module = "io.ktor:ktor-client-mock", version.ref = "ktor2" }
+
+ktor3-core = { module = "io.ktor:ktor-client-core", version.ref = "ktor3" }
+ktor3-engine-darwin = { module = "io.ktor:ktor-client-darwin", version.ref = "ktor3" }
+ktor3-engine-js = { module = "io.ktor:ktor-client-js", version.ref = "ktor3" }
+ktor3-mock = { module = "io.ktor:ktor-client-mock", version.ref = "ktor3" }
 
 okhttp-core = { module = "com.squareup.okhttp3:okhttp", version.ref = "okhttp" }
 okhttp-mockwebserver = { module = "com.squareup.okhttp3:mockwebserver", version.ref = "okhttp" }
