+++ settings.gradle.kts
@@ -18,7 +18,8 @@ include(
     "coil-compose",
     "coil-compose-core",
     "coil-network-core",
-    "coil-network-ktor",
+    "coil-network-ktor2",
+    "coil-network-ktor3",
     "coil-network-okhttp",
     "coil-gif",
     "coil-svg",
